confluence_search_agent/requirements.txt
google-adk[extensions]>=0.1.0
a2a-sdk>=0.3.0
httpx>=0.27.0
uvicorn>=0.27.0
fastapi>=0.109.0
pydantic>=2.0.0
litellm>=1.75.5
